reorganized tests, added tests for elementsFullHistory
